Set default encoding to utf8
This fixes #1. I played around a while trying to set `encoding='utf-8'`
in all the csv `reader` and `writer` constructors and `open` calls and
using the unicodecsv module. Nothing seemed to work.

I understand this is essentially a hack but it does work. I believe it's
a no-op in Python 3.

https://stackoverflow.com/questions/3828723/why-should-we-not-use-sys-setdefaultencodingutf-8-in-a-py-script